Create & modify README.md, index.html & tacos.html
Create and update index.html for Odin Recipes with a list of recipes as
well as a link to tacos.html which showcases a recipe for Chicken
Tacos. Create tacos.html webpage and update with an image of the
finished Chicken Tacos recipe, a brief description of the recipe, a list
of ingredients and directions on how to prepare.