Modify index.html & create pasta.html
Create pasta.html and include a recipe for Chicken Alfredo Pasta. Add
description of the recipe, ingredients needed and directions to prepare
the recipe. Add img/chicken_alfredo_pasta.png file to show a picture of
completed dish if directions are followed correctly. Modify index.html
to include a link that navigates to pasta.html in order to access the
Chicken Alfredo Pasta.